双ai对战

+++ game/src/main/java/com/hutcwp/game/wuziqi/GamePoint.kt
@@ -8,13 +8,13 @@ import androidx.annotation.ColorInt
  * YY: 909076244
  * type 白色球 ： 0  黑色球 ：1
  */
-class GamePoint(val x: Float, val y: Float, val type: Int, @ColorInt val color: Int) {
+class GamePoint(val x: Int, val y: Int, val type: Int, @ColorInt val color: Int) {
 
-    fun compare(x: Float, y: Float): Boolean {
+    fun compare(x: Int, y: Int): Boolean {
         return this.x == x && this.y == y
     }
 
-    fun isHaving(x: Float, y: Float, t: Int): Boolean {
+    fun isHaving(x: Int, y: Int, t: Int): Boolean {
         if (this.x == x && this.y == y) {
             if (type == t) return true
         }
